Move Async HTTP Client to its own namespace
