+++ composer.json
@@ -41,7 +41,7 @@
       "src/WordPress/Blueprints/functions.php",
       "src/WordPress/Zip/functions.php",
       "src/WordPress/Streams/stream_str_replace.php",
-      "src/WordPress/Streams/async_http_streams.php"
+      "src/WordPress/AsyncHttp/async_http_streams.php"
     ]
   },
   "scripts": {
